inicio do tratamento de exceções

+++ src/main/java/com/example/__projeto_springboot/service/ProdutoService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.example.__projeto_springboot.Repository.ProdutoRepository;
+import com.example.__projeto_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.__projeto_springboot.model.Produto;
 
 @Service
@@ -20,8 +21,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id) {
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id) {
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID"+id+"não encontrado."));
     }
 
     public Produto salvarProduto(Produto produto) {
@@ -29,6 +31,11 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id) {
+
+        if (!produtoRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto com ID"+id+"não encontrado.");
+        }
+
         produtoRepository.deleteById(id);
     }
 }
